Removed deprecated field nr_of_results
This was used in earlier versions to detect whether all responses had been received, but is deprecated in favor of using a "Completion" notification. When using older clients, which do not send this completion notification, this may result in AxonServer not being able to close a scatter-gather query before the indicated timeout. This is however considered normal operation and is merely lack of optimization.